+++ frontend/techpick/package.json
@@ -53,6 +53,7 @@
     "react-dom": "^18",
     "react-dropzone": "^14.3.5",
     "react-hot-toast": "^2.4.1",
+    "react-intersection-observer": "^9.15.1",
     "react-select": "^5.8.3",
     "react-spinners": "^0.14.1",
     "react-textarea-autosize": "^8.5.5",
